+++ src/components/business-card/index.tsx
@@ -1,138 +1,76 @@
-"use client"
+// ** React/Next.js Imports
+import React from "react";
+import Image from "next/image";
 
-import { useRef, useState } from "react"
-import { useFrame } from "@react-three/fiber"
-import { Text, useTexture } from "@react-three/drei"
-import type { Group } from "three"
+// ** Custom Components, Hooks, Utils, etc.
+import Card from "@/components/business-card/3DLicenseCard";
 
-interface BusinessCardProps {
-  rotation?: [number, number, number]
-}
+type CardData = {
+  name?: string;
+  title?: string;
+  subtitle?: string;
+  email?: string;
+  phone?: string;
+  location?: string;
+  image?: string;
+};
 
-export default function BusinessCard({ rotation = [0, 0, 0] }: BusinessCardProps) {
-  const cardRef = useRef<Group>(null)
-  const [hovered, setHovered] = useState(false)
-
-  // Use a placeholder texture for now - will be replaced with actual design
-  const texture = useTexture("/placeholder.svg")
-
-  useFrame((state) => {
-    if (cardRef.current) {
-      // Apply the rotation from props (controlled by scroll)
-      cardRef.current.rotation.x = rotation[0]
-      cardRef.current.rotation.y = rotation[1]
-      cardRef.current.rotation.z = rotation[2]
-
-      // Add a subtle floating animation
-      cardRef.current.position.y = Math.sin(state.clock.elapsedTime * 0.8) * 0.05
-    }
-  })
+const CardTemplate = ({ data }: { data?: CardData }) => {
 
   return (
-    <group
-      ref={cardRef}
-      onPointerOver={() => setHovered(true)}
-      onPointerOut={() => setHovered(false)}
-      scale={hovered ? 1.05 : 1}
-    >
-      {/* Card base with subtle gradient */}
-      <mesh castShadow receiveShadow position={[0, 0, 0]}>
-        <boxGeometry args={[3.5, 2, 0.05]} />
-        <meshStandardMaterial color="#ffffff" metalness={0.1} roughness={0.2} />
-      </mesh>
+    <>
+      <Card
+        style={{
+          width: "575px",
+          height: "350px",
+          backgroundColor: "blue",
+          backgroundSize: "cover",
+          backgroundPosition: "center",
+          padding: "1.5rem",
+          display: "flex",
+          justifyContent: "center",
+          alignItems: "center",
+          textAlign: "center",
+          transition: "background-image 0.4s ease-in-out",
+        }}
+      >
+        {/* Left side - Photo and Name */}
+        <div className="flex flex-col items-start justify-center text-black">
+          <div className="relative w-20 h-20 mb-3 rounded-full overflow-hidden">
+            <Image
+              src={data?.image || "/images/paul-3.png"}
+              alt={data?.name || "Paul McCartney"}
+              className="object-cover"
+              width={100}
+              height={100}
+            />
+          </div>
+          <h1 className="text-xl font-bold text-gray-800 mb-1">{data?.name || "Paul McCartney"}</h1>
+          <p className="text-sm text-gray-600 mb-1">{data?.title || "Musician & Songwriter"}</p>
+          <p className="text-xs text-gray-500">{data?.subtitle || "The Beatles • Wings"}</p>
+        </div>
 
-      {/* Front face content */}
-      <group position={[0, 0, 0.03]}>
-        <Text 
-          position={[-1.2, 0.6, 0]} 
-          fontSize={0.25}
-          color="#262626" 
-          anchorX="left"
-          font="/fonts/Inter-Bold.woff"
-          letterSpacing={0.02}
-        >
-          PAUL MCCARTNEY
-        </Text>
+        {/* Right side - Logo and Contact */}
+        <div className="flex flex-col items-end justify-between h-full">
+          <div className="relative w-24 h-12 mb-4">
+            <Image
+              src="/images/patchbay_black.png"
+              alt="Patchbay"
+              width={20}
+              height={20}
+              className="object-contain"
+            />
+          </div>
+          
+          <div className="text-right text-xs text-gray-600">
+            <p className="mb-1">{data?.email || "paul@patchbay.com"}</p>
+            <p className="mb-1">{data?.phone || "+44 20 7946 0958"}</p>
+            <p className="text-gray-500">{data?.location || "London, UK"}</p>
+          </div>
+        </div>
+      </Card>
+    </>
+  );
+};
 
-        <Text 
-          position={[-1.2, 0.3, 0]} 
-          fontSize={0.12} 
-          color="#404040" 
-          anchorX="left"
-          font="/fonts/Inter-Regular.woff"
-        >
-          Artist, Producer, Songwriter • BMI
-        </Text>
-
-        {/* Profile image */}
-        <mesh position={[1.2, 0.5, 0]} castShadow>
-          <circleGeometry args={[0.4, 32]} />
-          <meshStandardMaterial map={texture} />
-        </mesh>
-
-        {/* Dots in the corner */}
-        <group position={[1.2, 0, 0]}>
-          <mesh position={[-0.1, 0, 0]}>
-            <circleGeometry args={[0.05, 32]} />
-            <meshStandardMaterial color="#262626" />
-          </mesh>
-          <mesh position={[0.1, 0, 0]}>
-            <circleGeometry args={[0.05, 32]} />
-            <meshStandardMaterial color="#262626" />
-          </mesh>
-          <mesh position={[-0.1, -0.2, 0]}>
-            <circleGeometry args={[0.05, 32]} />
-            <meshStandardMaterial color="#262626" />
-          </mesh>
-          <mesh position={[0.1, -0.2, 0]}>
-            <circleGeometry args={[0.05, 32]} />
-            <meshStandardMaterial color="#262626" />
-          </mesh>
-        </group>
-
-        {/* Contract info */}
-        <Text 
-          position={[-1.2, -0.2, 0]} 
-          fontSize={0.12} 
-          color="#262626" 
-          anchorX="left" 
-          font="/fonts/Inter-Bold.woff"
-        >
-          CONTRACTING INFO
-        </Text>
-
-        <Text
-          position={[-1.2, -0.4, 0]}
-          fontSize={0.08}
-          color="#404040"
-          anchorX="left"
-          maxWidth={1.5}
-          font="/fonts/Inter-Regular.woff"
-        >
-          Paul McCartney 10 Downing St. London, UK SW1A2AA
-        </Text>
-
-        {/* Publishing info */}
-        <Text 
-          position={[0.5, -0.2, 0]} 
-          fontSize={0.12} 
-          color="#262626" 
-          anchorX="left" 
-          font="/fonts/Inter-Bold.woff"
-        >
-          PUB LINE
-        </Text>
-
-        <Text 
-          position={[0.5, -0.4, 0]} 
-          fontSize={0.08} 
-          color="#404040" 
-          anchorX="left"
-          font="/fonts/Inter-Regular.woff"
-        >
-          James Paul McCartney BMI IPI 12345678909
-        </Text>
-      </group>
-    </group>
-  )
-}
+export default CardTemplate;
